add send check code execute func

+++ src/main/java/vn/hanu/fit/ss2codecamp/controllers/TutorialController.java
@@ -33,7 +33,7 @@ public class TutorialController {
     }
 
     @GetMapping("/{id}")
-    public Tutorial getById(@PathVariable("id")Long id) throws TutorialNotFoundException {
+    public Tutorial getById(@PathVariable("id")int id) throws TutorialNotFoundException {
         return tutorialService.getById(id);
     }
